% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{analyze_processing_configs}
\alias{analyze_processing_configs}
\title{Analyse processing configurations for congruence.}
\usage{
analyze_processing_configs(data, best_method = FALSE)
}
\arguments{
\item{data}{data frame}

\item{best_method}{boolean (return only the best value (FALSE returns all))}
}
\value{
the best (or all) processing configurations and distance from the standard
}
\description{
Calculates the differences between fractional species abundances compared
 with the ZymoBIOMICS standard.
}
