% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{check_data}
\alias{check_data}
\title{Check user data for correct format.}
\usage{
check_data(data)
}
\arguments{
\item{data}{data frame}
}
\description{
Check that all specified columns are present, but additional columns are ignored.
}
