% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_network.R
\name{create_network_phyloseq}
\alias{create_network_phyloseq}
\title{Create network from a Phyloseq object.}
\usage{
create_network_phyloseq(
  physeq_object,
  max_dist = 0.5,
  distance_method = "bray"
)
}
\arguments{
\item{physeq_object}{Phyloseq object}

\item{max_dist}{integer (max distance between vertices)}

\item{distance_method}{string (see phyloseq documentation for available methods)}
}
\value{
list (CLI will save plot as image)
}
\description{
Creates a network showing co-expression relationships. More samples works best.
 For 'max_dist', lower is better. However:
 - Set higher if network too disconnected or contains few vertices
 - Set lower if network too connected
}
