% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa.R
\name{do_pcoa}
\alias{do_pcoa}
\title{Run PCoA analysis on data set.}
\usage{
do_pcoa(data, zero_missing = TRUE)
}
\arguments{
\item{data}{data frame}

\item{zero_missing}{TRUE/FALSE}
}
\value{
list which can be called to display the plot (CLI will save plot as image)
}
\description{
Creates a data frame of abundances for each sample against each species (mean),
 before multidimensional scaling. If a sample is missing abundances for a given
 species, the sample can either be removed (zero_missing = FALSE), or the specific
 entry can be set to zero (zero_missing = TRUE).
 
Removing samples is likely to be more destructive, particularly for lower quality data,
 and removing too many samples will reduce statistical weight.
}
