% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_user_data}
\alias{load_user_data}
\title{Read a single TSV file.}
\usage{
load_user_data(path)
}
\arguments{
\item{path}{path to single file as string}
}
\value{
a data frame with five columns (input columns plus 'entry_id' and 'source')
}
\description{
File should contain three columns (species, taxonomy, abundance) in any order, but is case-sensitive.
For loading multiple files at once, use load_user_data_dir().
}
