% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_user_data_dir}
\alias{load_user_data_dir}
\title{Read multiple TSV files from a directory.}
\usage{
load_user_data_dir(path)
}
\arguments{
\item{path}{path to a directory as string}
}
\value{
a data frame with five columns (input columns plus 'entry_id' and 'source')
}
\description{
Files should contain three columns (species, taxonomy, abundance) in any order, but is case-sensitive.
Directory should contain only TSV files to be loaded. Filenames will be used to create a 'source' column.
For loading a single file, use load_user_data().
}
