% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{make_barplot}
\alias{make_barplot}
\title{Make a simple bar plot of species abundance.}
\usage{
make_barplot(data, orientation = "vertical", max = 30)
}
\arguments{
\item{data}{data frame}

\item{orientation}{string (vertical/horizontal)}

\item{max}{integer (max number of entries to display (ordered by decreasing abundance),
any further entries are categorized into 'Other')}
}
\value{
ggplot object (CLI will save plot as image)
}
\description{
Creates a bar plot of mean abundance grouped by species.
}
