% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{make_clustered_heatmap}
\alias{make_clustered_heatmap}
\title{Create a heatmap of species abundance across sources.}
\usage{
make_clustered_heatmap(data)
}
\arguments{
\item{data}{data frame}
}
\value{
large HeatmapList (CLI will save plot as image)
}
\description{
Less clean than ggplot heatmap, but includes clustering and dendrograms
 for both samples (source) and species. Use 'make_heatmap()' for a simpler
 but cleaner-looking heatmap.
}
