% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{make_density_plot}
\alias{make_density_plot}
\title{Make density plot.}
\usage{
make_density_plot(data)
}
\description{
Each file in directory (or each different source) plotted separately on
 the same graph.
 
 @param data data frame
 @returns ggplot object (CLI will save plot as image)
}
