% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Create a heatmap of species abundance across sources.}
\usage{
make_heatmap(data)
}
\arguments{
\item{data}{data frame}
}
\value{
ggplot object (CLI will save plot as image)
}
\description{
Makes a clean heatmap but does not include clustering or dendrograms.
Use 'make_clustered_heatmap()' for these functions.
}
