% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{make_treemap}
\alias{make_treemap}
\title{Make a tree map to show most abundant species.}
\usage{
make_treemap(data, max)
}
\arguments{
\item{data}{data frame}

\item{max}{integer (takes top n species)}
}
\value{
ggplot object (CLI will save plot as image)
}
\description{
Samples are grouped (if applicable) to show top mean abundances.
}
