% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{plot_controls}
\alias{plot_controls}
\title{Plot fractional species abundance against Zymo standard.}
\usage{
plot_controls(data)
}
\arguments{
\item{data}{data frame}
}
\value{
ggplot object (CLI will save plot as image)
}
\description{
Plot fractional species abundance against Zymo standard.
}
