% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_network.R
\name{create_physeq_object}
\alias{create_physeq_object}
\title{Create a Phyloseq object from data set.}
\usage{
create_physeq_object(data)
}
\arguments{
\item{data}{data frame}
}
\value{
Phyloseq object
}
\description{
Required to plot networks.
}
\details{
How this object is created depends on how networks are created later on.
 It is important to arrange the input data frame in the same way to ensure
 that networks are both reproducible and comparable.
}
